requirements.txt
opencv-python==4.8.0.74
numpy==1.25.2
scikit-image==0.19.3
matplotlib==3.7.1
Pillow